New file structure & server.js update
- updated the server.js (previously index.js) to use app.static instead of serving each file manually.

- Removed the image from the html and am loading it on the code.

- Messing with the animation function, but will leave this for later since that is much more to do.

Btw, I'm making this update to be able to showcase this project on the web, making it easier to explore such a fun project.

And I'll be honest, I'm quite impressed by how high quality the final results are, since I made this on my first ear of programming. Way to go, little me!