Star drawing optimized
- Optimized the drawing of the background stars by only drawing the stars that are on the screen;

I do not like the way drawings optimizations are done in classes that should not be doing it. After I implement the parallax and animation to the panzoom class, I will add a system to make only one class be responsible for drawing and optimizing the drawing (such as a "camera" class).

I'm still figuring this out with the voices on my head. They are already having some ideas, so I'll just let them work.